fix: fixes comments related implementations
It turned out how to save comments was not simple. A way to get a single
post is done in another component. The post is passed to Post component.
While comments of that post are mounted to the Post component. Just an
array type for comments overrides the data by the later post.

For that reason, comments are saved in a map whose keys are parent id.
It is a big change, so this commit is a big one as well.